### 2020-05-28 #381 @eribul
coder: Deterministic Categorization of Items Based on External Code Data
Submitting Author: Erik Bülow (@eribul)
Repository:  https://github.com/eribul/coder
Version submitted:   0.11.9
Version accepted: TBD

---

```
Package: coder
Title: Deterministic Categorization of Items Based on External Code Data
Version: 0.11.9
Authors@R: person("Erik", "Bulow", email = "eriklgb@gmail.com",
                  comment = c(ORCID = "0000-0002-9973-456X"))
  Fast categorization of items based on external code data identified by
  regular expressions.
  It is easy to introduce new classification schemes (by 'classcodes' objects) or
  to use default schemes included in the package. Use cases includes patient
  categorization based on co-morbidity indices such as Charlson, Elixhauser,
  RxRisk V, or the co-morbidity-polypharmacy score (CPS), as well as adverse
  events after arthroplasty surgery.
  covr,
  writexl
  decoder
  knitr
```

	- [ ] workflow automataion

The aim of the package is to categorize items/individuals/patients from one large data sets, with the help of external code data using a dynamic system of classicifaction schemes based on regular expression. There are no analytical/statistical methods included.

The primary user group might be medical/epidemiological researchers classifying patients by medical scores such as comorbidity indices or adverse events. The package might be used more broadly however due to generic principles.

There are some R packages with a narrow focus on Charlson and Elixhauser co-morbidity based on ICD-9/10-codes (`icd`, `comorbidity`, `medicalrisk`, `comorbidities.icd10` and  `icdcoder`).
`icd` and `comorbidity` are both good for their purpose. `medicalrisk` can be used with ICD-9-CM codes but is not up-to-date with the latest version of ICD-10. `comorbidities.icd10` and `icdcoder` are not actively developed or maintained.
The `coder` package provides greater flexibility for combining different sets of codes (ICD-8, ICD-9, ICD-9-CM and ICD-10 etc as given by regular expressions, either as included in the package by default, or as provided by the user). All other packages relies on pre-specified (hard-coded) classifications. None of the other packages includes any other classification then Charlson/Elixhauser (such as the RxRisk V classification based on medical ATC codes, which is  also included in the `coder` package by default).
