### 2020-06-04 #381 @annakrystalli
Hi @eribul and thanks for your submission!

After discussion, the editors have found your package to be in scope. I'll be assigning a handling editor shortly 👍
