### 2020-06-11 #381 @eribul
Thank you very much @noamross! I am very happy for your response and will just briefly adress some of your comments:

I was actually equaliy amazed regarding the name :-) I do like `coder` however since: 1) I have another CRAN package called `decoder` so they fit quite nicely together. 2) I realise there is some ambiguity but that is kind of intential since both "medical coding" and "computer coding" is relevant :-) 3) I did use another name initially, that was later used by others, 4) As you mention, the package has already been used for some published work.

Considering DOI, I would be happy to provide such trough Zenodo. Just let me know, and I will fix that!

I realize that the importing of packages is unusal but one of them is `data.table` (for which this is recommended), and the other one is my other package `decoder` that is also intermingled with this package in such a way that this seems reasonable. I have been careful to not depend on any other packages (although a few more are suggested, mostly for developmnent purposes).

I am looking forward to the review process and thank you once more!
