### 2020-06-11 #381 @noamross
Great, @eribul.  You don't need a DOI until the end of the review process to transfer the review to JOSS.

We have two reviewers: @zabore and @dgrtwo, thank you for volunteering!  Due date is 2020-07-02.  Please note that this is a pass-through review to JOSS, so please include in your review whether `paper/paper.md` conforms to [JOSS requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
