### 2020-07-13 #381 @noamross
Thanks for your detailed review, @dgrtwo! @eribul, lots to digest in both of these!  Please let us know if you have questions and update us with a summary of changes in your response.
