### 2020-07-14 #381 @eribul
Dear @noamross @dgrtwo and @annakrystalli I am so impressed by your efforts and comments and I am really looking forward to improve the package accordingly! I am currently on vacation without computer access however. I therefore hope that a slight delay would be acceptable? Thank you once more, I really enjoy this experience!
