### 2020-07-15 #381 @noamross
@eribul, yes, that's fine. We aim for a _response_ to reviews within two weeks, even if all the changes involved are yet to be completed, but think of that timeframe as starting when you return.  Enjoy your vacation!
