### 2020-08-03 a1c1a81
Uninformative test headings. Fix #109.
Review: Thank you for the notice! Those were the default names from usethis::use_test(). I have corrected accordingly!
