+++ man/elixhauser.Rd
@@ -59,18 +59,18 @@ Elixhauser A, Steiner C, Harris DR, Coffey RM (1998).
 Moore, B. J., White, S., Washington, R., Coenen, N., & Elixhauser, A. (2017).
   Identifying Increased Risk of Readmission and In-hospital Mortality Using
   Hospital Administrative Data.
-  Medical Care, 55(7), 698–705. http://doi.org/10.1097/MLR.0000000000000735
+  Medical Care, 55(7), 698–705. \url{http://doi.org/10.1097/MLR.0000000000000735}
 
 Quan Hude et al. (2005). Coding algorithms for defining
   comorbidities in ICD-9-CM and ICD-10 administrative data.
   Medical care, 1130-1139.
-  \url{www.jstor.org/stable/3768193}
+  \url{https://www.jstor.org/stable/3768193}
 
 Thompson, N. R., Fan, Y., Dalton, J. E., Jehi, L., Rosenbaum, B. P.,
   Vadera, S., & Griffith, S. D. (2015).
   A new Elixhauser-based comorbidity summary measure to predict in-hospital
   mortality. Med Care, 53(4), 374–379.
-  http://doi.org/10.1097/MLR.0000000000000326
+  \url{http://doi.org/10.1097/MLR.0000000000000326}
 
 Walraven, C. Van, Austin, P. C., Jennings, A., Quan, H., Alan, J., Walraven,
   C. Van, … Jennings, A. (2009).
