+++ DESCRIPTION
@@ -25,7 +25,7 @@ Imports:
   data.table,
   decoder
 LazyData: TRUE
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 VignetteBuilder: 
   knitr
 URL: https://github.com/eribul/coder
