### 2020-08-03 3d26e7e
vignette headings same as engine name entries
