+++ vignettes/Comorbidity.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Co-morbidity"
+title: "Comorbidity"
 author: "Erik Bulow"
 date: "`r Sys.Date()`"
 output: 
