+++ .github/workflows/pkgdown.yaml
@@ -37,6 +37,12 @@ jobs:
       - name: Install package
         run: R CMD INSTALL .
 
+      - name: Reveal identity
+        run: |
+          git config --global user.email "eriklgb@gmail.com"
+          git config --global user.name "Erik Bülow"
+
+
       - name: Deploy package
         run: pkgdown::deploy_to_branch(new_process = FALSE)
         shell: Rscript {0}
