### 2020-08-03 2619d32
changed check versions
