+++ R/set_classcodes.R
@@ -5,9 +5,11 @@
 #' @param regex name of column with regular expressions to use for
 #'   classification, either with or without prefix \code{regex_}.
 #'   \code{NULL} (default) uses the first classcodes column with prefix "regex".
-#'   This should be a sensible choice for default classcodes object included in the package.
-#' @param start,stop should codes start/end with the specified regular expressions?
-#'   If \code{TRUE}, column "regex" is prefixed/suffixed by "^"/"$".
+#'   This should be a sensible choice for default classcodes object included
+#'   in the package.
+#' @param start,stop should codes start/end with the specified regular
+#'   expressions? If \code{TRUE}, column "regex" is prefixed/suffixed
+#'   by "^"/"$".
 #' @param tech_names should technical column names be used? If \code{FALSE},
 #'   colnames are taken directly from group names of \code{by}, if \code{TRUE},
 #'   these are changed to more technical names avoiding special characters and
