### 2020-08-03 #381 @eribul
Once again I must appologize for my late response! I have been back from vacation but had to put intense work into my PhD thesis with deadline from the print shop today. I am now on vacation once more, for an additional two weeks. I do have access to my computer, however, and I hope to be able to do at least some work :-)

I have organized the raied issues from @zabore here: https://github.com/eribul/coder/issues/89
and from @dgrtwo here https://github.com/eribul/coder/issues/90
I have also fixed some of the smaller issues from @zabore

It is my plan to go through all the issues and then to summarize my response here.
