### 2020-08-10 6be4d59
Remove toy example with cars. Fix #93.
