+++ man/ex_cars.Rd
@@ -1,28 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/manual_for_datasets.R
-\docType{data}
-\name{ex_cars}
-\alias{ex_cars}
-\title{Example data for can registry}
-\format{
-Data frames with 1,000 rows and 3 variables:
-\describe{
-  \item{id}{random person names}
-  \item{code}{random car models}
-  \item{code_date}{random dates for car ownership}
-}
-}
-\usage{
-ex_cars
-}
-\description{
-Example data for fictive people and their car ownership.
-}
-\seealso{
-Other example data: 
-\code{\link{ex_carbrands}},
-\code{\link{ex_icd10}},
-\code{\link{ex_people}}
-}
-\concept{example data}
-\keyword{datasets}
