+++ man/ex_carbrands.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/manual_for_datasets.R
-\docType{data}
-\name{ex_carbrands}
-\alias{ex_carbrands}
-\title{Example classcodes object for classification of car brands by their producer}
-\format{
-Data frames with 11 rows and 2 variables:
-\describe{
-  \item{group}{Car makers}
-  \item{regex}{Regular expressions identifying car brands}
-}
-}
-\usage{
-ex_carbrands
-}
-\description{
-Example classcodes object for classification of car brands by their producer
-}
-\seealso{
-Other example data: 
-\code{\link{ex_cars}},
-\code{\link{ex_icd10}},
-\code{\link{ex_people}}
-}
-\concept{example data}
-\keyword{datasets}
