### 2020-08-10 e63cd7b
Use case in README. Fix #91.
