+++ vignettes/coder.Rmd
@@ -18,6 +18,8 @@ knitr::opts_chunk$set(
 ```
 
 
+A typical use case of the `coder` package would consider patient data and comorbidity as described in the package [readme](https://github.com/eribul/coder/).
+
 The concept of comorbidity is often attributed to [Feinstein](https://doi.org/10.1016/0021-9681(70)90054-8):
 
 > [T]he term co-morbidity will refer to any distinct additional clinical entity that has existed or that may occur during the clinical course of a patient who has the index disease under study. 
