+++ tests/testthat/test-classify.R
@@ -49,48 +49,3 @@ test_that("df/dt", {
   expect_equal(ncol(cl) + 1, ncol(as.data.frame(cl)))
   expect_equal(ncol(cl) + 1, ncol(as.data.table(cl)))
 })
-
-
-# condition ---------------------------------------------------------------
-
-set.seed(123)
-
-people <-
-  data.frame(
-    name = c(
-      "Medina, Desmand", "al-Hashem,  Muzna",
-      "el-Kaleel, Aadil", "Pettsson, Svenne"
-    ),
-    license = as.Date(c(
-      "1949-05-23", "2016-03-04", "2019-12-18", "1985-05-04"
-    )),
-    cond1 = TRUE,
-    cond2 = FALSE,
-    cond3 = NA,
-    stringsAsFactors = FALSE
-  )
-
-cd <- codify(people, ex_cars, id = "name")
-
-# Use diff
-test_that("condition", {
-
-  ex_carbrands$condition <- "cond1"
-  expect_equal(
-    sum(suppressWarnings(classify(cd, ex_carbrands)), na.rm = TRUE),
-    5
-  )
-
-  ex_carbrands$condition <- "cond2"
-  expect_equal(
-    sum(suppressWarnings(classify(cd, ex_carbrands)), na.rm = TRUE),
-    0
-  )
-
-  ex_carbrands$condition <- "cond3"
-  expect_true(
-    all(suppressWarnings(classify(cd, ex_carbrands)) %in% c(NA, FALSE))
-  )
-})
-
-
