### 2020-08-10 0fba8b2
Better documentation of default indices. Fix #94.
