+++ R/categorize.R
@@ -12,7 +12,8 @@
 #'   passed to argument
 #'   \code{by} in function \code{\link{index}}. It is possible to
 #'   include several indices as a character vector.
-#'   \code{NULL} will include all available indices.
+#'   \code{NULL} will include all available indices as specified by the
+#'   "indices" attribute of the \code{cc} object (\code{attr(cc, "indices")})
 #' @param sort logical. Should output be sorted by the 'id' column?
 #'   (This could effect computational speed for large data sets.)
 #'   Data is sorted by 'id' internally. It is therefore faster to keep the
@@ -22,10 +23,10 @@
 #' @param codify_args List of named arguments passed to \code{\link{codify}}
 #' @inheritParams classify
 #'
-#' @return data frame ('data.table' object) with all data from
-#' the input data set \code{to/data} combined with logical columns indicating
-#' membership of categories identified by the classcode object.
-#' Indices are also included if so specified by the 'index' argument.
+#' @return Object of class \code{data.table} made from \code{data} combined with
+#' logical columns indicating membership of categories identified by the
+#' \code{classcodes} object (the \code{cc} argument).
+#' Indices are also included if specified by the 'index' argument.
 #'
 #' @export
 #'
@@ -33,7 +34,7 @@
 #' # Add Elixhauser based on all registered ICD10-codes
 #' categorize(ex_people, ex_icd10, "elixhauser", id = "name")
 #'
-#' # Add Charlson categorias and two versions of a calculated index.
+#' # Add Charlson categories and two versions of a calculated index.
 #' # Only include recent hospital visits within 30 days before surgery,
 #' # and use technical variable names to clearly identify the new columns.
 #' categorize(ex_people, ex_icd10, "charlson",
