### 2020-08-10 b0fcb4c
Visualize visualize(). Fix #95.
