### 2020-08-17 05fe4e1
Triad description in DESRIPTION. #102.
