### 2020-08-17 1642f52
Explain triad. #102.
