+++ vignettes/coder.Rmd
@@ -17,12 +17,17 @@ knitr::opts_chunk$set(
 )
 ```
 
+# Introduction
 
-A typical use case of the `coder` package would consider patient data and comorbidity as described in the package [readme](https://eribul.github.io/coder/).
+## Triad of objects
 
-The concept of comorbidity is often attributed to [Feinstein](https://doi.org/10.1016/0021-9681(70)90054-8):
+Functions of the package relies on a triad of objects: 
 
-> [T]he term co-morbidity will refer to any distinct additional clinical entity that has existed or that may occur during the clinical course of a patient who has the index disease under study. 
+1. Case data with unit id:s and possible dates of interest
+2. External code data for corresponding units in (1) and with optional dates of interest and
+3. A classification scheme ('classcodes' object) with regular expressions to identify and categorize relevant codes from (2). 
+
+It is easy to introduce new classification schemes ('classcodes' objects) or to use default schemes included in the package. Use cases includes patient categorization based on comorbidity indices such as Charlson, Elixhauser, RxRisk V, or the co-morbidity-polypharmacy score (CPS), as well as adverse events after hip and knee arthroplasty surgery.
 
 The `coder` package contains some default classcodes objects for co-morbidity:
 
@@ -35,10 +40,29 @@ knitr::kable(
   cl[cl$classcodes %in% c("elixhauser", "charlson", "cps", "rxriskv"), ]
 )
 ```
+## Triad of functions
+
+There are three important functions to control the intended work flow of the package:
+
+i. `codify()` will merge object (1) and (2) for a coded data set of the intended format. If optional dates are specified, those will be used to construct time windows in order to filter out only the important dates (i.e. comorbidity during one year before surgery or adverse events 90 days after).
+ii. `classify()` will then use the coded data and classify it using the `classcodes` object (3) (i.e. to code comorbidity data by the Charlson or Elixhauser comorbidity classifications, see below).
+iii. `index()` is a third optional step to summarize the individual `classcodes` categories to a (possibly weighted) index sum for each coded item (i.e. to calculate the Charlson comorbidity index for each patient).
+
+Those steps could be performed explicitly as `codify() %>% classify() %>% index()` or implicitly by the main function `categorize()` combining all steps automatically.  
+
+
+# Comorbidity use case
+
+A typical use case of the `coder` package would consider patient data and comorbidity as described in the package [readme](https://eribul.github.io/coder/).
+
+The concept of comorbidity is often attributed to [Feinstein](https://doi.org/10.1016/0021-9681(70)90054-8):
+
+> [T]he term co-morbidity will refer to any distinct additional clinical entity that has existed or that may occur during the clinical course of a patient who has the index disease under study. 
+
 
 # Charlson and Elixhauser
 
-The [Charlson](https://doi.org/10.1016/0021-9681(87)90171-8) and [Elixhauser](https://www.jstor.org/stable/3766985) co-morbidity indices are two  examples used in medical research. Each index consist of several medical conditions, possibly summarized by a (weighted) index. Each condition is defined by a set of medical codes, such as described by [Quan et al.](https://dx.doi.org/10.1097/01.mlr.0000182534.19832.83) and many others. Different versions of the International Classification of Diseases (ICD) codes are often used.
[…]
