### 2020-08-17 dacbaf7
Explain triad in categorize help. Fix #102.
