+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method("[",classcodes)
 S3method("[<-",classcodes)
 S3method(as.data.frame,classified)
 S3method(as.data.table,classified)
+S3method(as.keyvalue,classcodes)
 S3method(classify,data.frame)
 S3method(classify,default)
 S3method(filter_dates,Date)
@@ -16,7 +17,6 @@ S3method(summary,classcodes)
 export(all_classcodes)
 export(as.classcodes)
 export(as.codedata)
-export(as.keyvalue.classcodes)
 export(categorize)
 export(classify)
 export(codebook)
