### 2020-08-17 0a3a090
improved vignette. Fix #99.
