+++ vignettes/coder.Rmd
@@ -19,6 +19,8 @@ knitr::opts_chunk$set(
 
 # Introduction
 
+The `coder` package simplifies unit classification based on external code data. this is an abstract aim that is both generic, but which might also be hard to grasp without further concretization. In this vignette, I will first explain the overall design principles, and then exemplify the concept with a typical use case involving patients with total hip arthroplasty and their pre-surgery comorbidity. Note, however, that the package is not limited to patient data or medical settings, although this use case constituted the initial motivation for the package.
+
 ## Triad of objects
 
 Functions of the package relies on a triad of objects: 
@@ -34,6 +36,7 @@ The `coder` package contains some default classcodes objects for co-morbidity:
 ```{r, warning=FALSE}
 
 library(coder)
+library(tibble)
 
 cl <- coder::all_classcodes()
 knitr::kable(
@@ -51,13 +54,22 @@ iii. `index()` is a third optional step to summarize the individual `classcodes`
 Those steps could be performed explicitly as `codify() %>% classify() %>% index()` or implicitly by the main function `categorize()` combining all steps automatically.  
 
 
-# Comorbidity use case
+# Use case
 
 A typical use case of the `coder` package would consider patient data and comorbidity as described in the package [readme](https://eribul.github.io/coder/).
 
 The concept of comorbidity is often attributed to [Feinstein](https://doi.org/10.1016/0021-9681(70)90054-8):
 
 > [T]he term co-morbidity will refer to any distinct additional clinical entity that has existed or that may occur during the clinical course of a patient who has the index disease under study. 
+
+Let's consider a group of patients with total hip arthroplasty (THA), as identified from a national quality register, which might be large in size. Assume we are interested in those patients pre-surgery comorbidity, which is not captured by the quality register itself. Instead, this data might be codified in a secondary source, such as a national patient register containing all hospital visits and admissions during several years, both before and after the THA-surgery. Each hospital visit/admission might be recorded with one or several medical codes, for example using the International classification of diseases version 10 (ICD-10). Similarly, a medical prescription register might hold records of prescribed drugs with their corresponding codes from the Anatomic therapeutic chemical classification (ATC) system. 
+
+Thus, combining the primary and secondary data sets (objects 1-2 above) using some unique patient id, and a possible time window (i.e. to only consider comorbidity as recorded during one year before the THA), is a first step to identify patient comorbidity. This step is performed by the `codify()` function in step (i) above.
+
+We have now gathered all the relevant codes for each patient. Common classifications (i.e. ICD-10 and ATC) are wast, however, including tens of thousands of medical/chemical codes, which are cumbersome and impractical to use directly. It is therefore common to categorize such codes into broader categories (i.e. by the Charlson, Elixhauser or RxRisk V classifications as below). Such classification could be a simple code matching problem using a look-up table. This is generally a slow, cumbersome and error-prone process, however. I therefore recommend to use regular expression for a compact code representation, as well as a computationally faster procedure. This is implemented in the `classify()` function from step (ii) above.
+
[…]
