### 2020-08-17 2ebb230
SPecify default dates. Fix #105.
