+++ man/set_classcodes.Rd
@@ -39,6 +39,14 @@ are prefixed by the name of the classification scheme.}
 \description{
 Set classcodes object
 }
+\examples{
+# Prepare a classcodes object for the Charlson comorbidity classification
+# based on the default regular expressions
+set_classcodes(charlson)
+
+# Same as above but based on regular expressions for ICD-8
+set_classcodes(charlson, regex = "regex_icd8_brusselaers")
+}
 \seealso{
 Other classcodes: 
 \code{\link{all_classcodes}()},
