### 2020-08-17 ef36ff0
Include example. Fix #106.
