+++ man/as.keyvalue.classcodes.Rd
@@ -22,7 +22,7 @@ S3-method for generic \code{\link[decoder:keyvalue]{decoder::as.keyvalue()}}
 # List all codes with corresponding classes as recognized by the Elixhauser
 # comorbidity classification according to the Swedish version of the
 # international classification of diseases version 10 (ICD-10-SE)
-as.keyvalue(elixhauser, "icd10se")
+decoder::as.keyvalue(elixhauser, "icd10se")
 }
 \seealso{
 Other helper: 
