### 2020-08-20 9c52b7f
Make categorize generic. Fix #111.
