+++ tests/testthat/test-categorize.R
@@ -1,3 +1,8 @@
+
+x <- codify(ex_people, ex_icd10, id = "name",
+            date = "event", days = c(-365, 0))
+
+
 test_that("categorize", {
   expect_message(
     categorize(ex_people, ex_icd10, "elixhauser", id = "name"),
@@ -11,8 +16,8 @@ test_that("categorize", {
       categorize(ex_people[c(1, 1), ], ex_icd10, "elixhauser", id = "name"),
       "Non-unique ids!"
     )
-  )
 
+  )
   expect_equal(
     ncol(categorize(ex_people, ex_icd10, "elixhauser",
                     id = "name", index = "index_sid30")),
@@ -35,5 +40,15 @@ test_that("categorize", {
   expect_equal(
     nrow(categorize(ex_people, ex_icd10, "elixhauser", id = "name")),
     nrow(ex_people)
+  )
+
+  expect_warning(
+    categorize(x, "charlson", id = "name"),
+    "Output might contain extra columns as left-overs"
+  )
+
+  expect_identical(
+    suppressWarnings(categorize(x, "charlson", id = "name")[,-"hdia"]),
+    categorize(ex_people, ex_icd10, "charlson", id = "name", codify_args = list(date = "event", days = c(-365, 0)))
   )
 })
