+++ NAMESPACE
@@ -6,6 +6,9 @@ S3method("[<-",classcodes)
 S3method(as.data.frame,classified)
 S3method(as.data.table,classified)
 S3method(as.keyvalue,classcodes)
+S3method(categorize,codified)
+S3method(categorize,data.frame)
+S3method(categorize,data.table)
 S3method(classify,data.frame)
 S3method(classify,default)
 S3method(filter_dates,Date)
