### 2020-08-21 84ab48f
Make visualize S3-method of generics::visualize.
Fix #123 .
