+++ R/reexports.R
@@ -1,0 +1,7 @@
+#' @importFrom tibble %>%
+#' @export
+tibble::`%>%`
+
+#' @importFrom generics visualize
+#' @export
+generics::visualize
