+++ R/visualize.R
@@ -1,4 +1,3 @@
-
 #' Visualize classification scheme in web browser
 #'
 #' Classes are visualized by their regular expressions in a web browser.
@@ -6,6 +5,8 @@
 #' weights but might be useful both for understanding of a classification scheme
 #' in use, and during the creation and debugging of such.
 #'
+#' @param x [classcodes] object or name of such object included in the package
+#' (see [all_classcodes()]).
 #' @inheritParams set_classcodes
 #' @param group names (as character vector) of groups to visualize
 #' (all groups if `NULL`)
@@ -40,14 +41,19 @@
 #'  # Get URL for later visualization
 #'  visualize("hip_ae", show = FALSE)
 #' @family classcodes
-visualize <- function(cc, group = NULL, show = TRUE, ...) {
-  cc <- set_classcodes(cc, ...)
-
+#' @importFrom generics visualize
+visualize.classcodes <- function(x, group = NULL, show = TRUE, ...) {
   if (!is.null(group))
-    cc <- cc[cc$group %in% group, ]
-  r <- paste0(cc$regex, collapse = "|")
+    x <- x[x$group %in% group, ]
+  r <- paste0(x$regex, collapse = "|")
   r <- utils::URLencode(r)
   u <- paste0("https://jex.im/regulex/#!embed=true&flags=&re=", r)
   if (show) utils::browseURL(u)
   invisible(u)
+}
+
+#' @export
+visualize.character <- function(x, group = NULL, show = NULL, ...) {
+  x <- set_classcodes(x, ...)
+  visualize(x, ...)
 }
