+++ man/classcodes.Rd
@@ -80,7 +80,7 @@ Other classcodes:
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
-\code{\link{visualize}()}
+\code{\link{visualize.classcodes}()}
 
 Other classcodes: 
 \code{\link{all_classcodes}()},
@@ -90,6 +90,6 @@ Other classcodes:
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
-\code{\link{visualize}()}
+\code{\link{visualize.classcodes}()}
 }
 \concept{classcodes}
