+++ man/visualize.classcodes.Rd
@@ -1,13 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/visualize.R
-\name{visualize}
-\alias{visualize}
+\name{visualize.classcodes}
+\alias{visualize.classcodes}
 \title{Visualize classification scheme in web browser}
 \usage{
-visualize(cc, group = NULL, show = TRUE, ...)
+\method{visualize}{classcodes}(x, group = NULL, show = TRUE, ...)
 }
 \arguments{
-\item{cc}{\code{\link{classcodes}} object or name of such}
+\item{x}{\link{classcodes} object or name of such object included in the package
+(see \code{\link[=all_classcodes]{all_classcodes()}}).}
 
 \item{group}{names (as character vector) of groups to visualize
 (all groups if \code{NULL})}
