+++ man/print.summary.classcodes.Rd
@@ -31,6 +31,6 @@ Other classcodes:
 \code{\link{codebook}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
-\code{\link{visualize}()}
+\code{\link{visualize.classcodes}()}
 }
 \concept{classcodes}
