+++ man/reexports.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/reexports.R
+\docType{import}
+\name{reexports}
+\alias{reexports}
+\alias{\%>\%}
+\alias{visualize}
+\title{Objects exported from other packages}
+\keyword{internal}
+\description{
+These objects are imported from other packages. Follow the links
+below to see their documentation.
+
+\describe{
+  \item{generics}{\code{\link[generics]{visualize}}}
+
+  \item{tibble}{\code{\link[tibble:reexports]{\%>\%}}}
+}}
+
