### 2020-08-23 152d445
Make date functions internal. Fix #119.
