+++ R/codedata.R
@@ -1,12 +1,19 @@
 #' Code data
 #'
-#' Make structured code data set based on [data.frame].
+#' Make structured code data set.
 #'
-#' @inheritDotParams dates_within to from
-#' @param x [data.frame] with columns `id` (character), `code`,
+#' @param x data table/frame with columns `id` (character), `code`,
 #'   and optional `code_date` (Date).
+#' @param id,code column names (from `x`) with case id:s (character) and
+#'    codes of interest.
+#' @param code_date optional column name (from `x`) with relevant dates
+#'  corresponding to each code. If there is a column named `code_date` this will
+#'  be used by default.
+#' @param period optional date vector of length two with lower and upper limits
+#'   for inclusion period
 #' @param .setkeys should output be indexed (Beneficial for large data sets.)
 #' @param alnum Should codes be cleaned from all non alphanumeric characters?
+#' @param ... arguments passed between methods
 #'
 #' @return
 #'   [data.table] with columns: `id` (character), `code` and
@@ -19,13 +26,13 @@
 #' x <-
 #'   data.frame(
 #'     id               = "1",
-#'     code_date        = as.Date("2017-02-02"),
+#'     code_date        = as.Date("2017-01-01"),
 #'     code             = "a",
 #'     stringsAsFactors = FALSE
 #'   )
 #' as.codedata(x)
 #'
-#' # Drop dates outside specified limits
+#' # Ignore codes for dates outside specified limits
 #' y <-
 #'   data.frame(
 #'     id               = "2",
@@ -34,24 +41,44 @@
 #'     stringsAsFactors = FALSE
 #'   )
 #' z <- rbind(x, y)
-#' as.codedata(z)
+#' as.codedata(z, period = c(as.Date("1970-01-01"), Sys.Date()))
 #'
 #' @family codedata
-as.codedata <- function(x, ..., .setkeys = TRUE, alnum = FALSE) {
-
-  code_date <- id <- code <- NULL # Fix for R Check
-  hasdates  <- "code_date" %in% names(x)
-  keys      <- c("id", if (hasdates) "code_date", "code")
-
-  check_codedata(x)
-
-  if (!is.data.table(x)) x <- as.data.table(x)
-  if (hasdates)          x <- x[dates_within(code_date, ...)]
-  if (alnum)             x[, code := gsub("[^[:alnum:]]", "", code)]
-  if (.setkeys)          setkeyv(x, keys)
-
-  unique(x, by = keys)[]
+#' @name as.codedata
+as.codedata <- function(x, ...) {
+  UseMethod("as.codedata")
 }
+
+
+#' @export
+as.codedata.data.frame <- function(x, ...) {
+  as.codedata(as.data.table(x), ...)
+}
+
+#' @export
+#' @rdname as.codedata
+as.codedata.data.table <-
+  function(
+    x, id = "id", code = "code",
+    code_date = if (utils::hasName(x, "code_date")) "code_date" else NULL,
+    period = NULL, .setkeys = TRUE, alnum = FALSE, ...) {
+
+    setnames(x, c(id, code, code_date),
+             c("id", "code", if (!is.null(code_date)) "code_date"))
+
+    code_date <- id <- code <- NULL # Fix for R Check
+    hasdates  <- utils::hasName(x, "code_date")
+    keys      <- c("id", if (hasdates) "code_date", "code")
+
+    check_codedata(x)
+
+    if (hasdates & !is.null(period))
+      x <- x[dates_within(code_date, from = period[1], to = period[2])]
+    if (alnum)    x[, code := gsub("[^[:alnum:]]", "", code)]
+    if (.setkeys) setkeyv(x, keys)
+
+    unique(x, by = keys)[]
+  }
 
 
[…]
