+++ NAMESPACE
@@ -3,6 +3,8 @@
 S3method("$<-",classcodes)
 S3method("[",classcodes)
 S3method("[<-",classcodes)
+S3method(as.codedata,data.frame)
+S3method(as.codedata,data.table)
 S3method(as.data.frame,classified)
 S3method(as.data.table,classified)
 S3method(as.keyvalue,classcodes)
@@ -28,8 +30,6 @@ export(classify)
 export(codebook)
 export(codebooks)
 export(codify)
-export(dates_within)
-export(filter_dates)
 export(index)
 export(is.classcodes)
 export(is.codedata)

[…]
