### 2020-09-23 76e966d
Update visualize.R
Fix #127
