### 2020-09-24 9223f70
return tibble frmo all_classcodes
Part of #121
