+++ NAMESPACE
@@ -34,7 +34,6 @@ export(index)
 export(is.classcodes)
 export(is.codedata)
 export(set_classcodes)
-export(spss2date)
 export(visualize)
 import(data.table)
 importFrom(decoder,as.keyvalue)
