### 2020-09-24 7bd8ed7
refactoring codedata.
FIx #117.
