+++ tests/testthat/test-categorize.R
@@ -1,11 +1,12 @@
 
-x <- codify(ex_people, ex_icd10, id = "name",
-            date = "event", days = c(-365, 0))
+
+x <- codify(ex_people, codedata = ex_icd10, id = "name", code = "icd10",
+            date = "surgery", code_date = "admission", days = c(-365, 0))
 
 
 test_that("categorize", {
   expect_message(
-    categorize(ex_people, ex_icd10, "elixhauser", id = "name"),
+    categorize(ex_people, codedata = ex_icd10, cc = "elixhauser", id = "name", code = "icd10"),
     "Classification based on: regex_icd10"
   )
   expect_error(
@@ -13,42 +14,57 @@ test_that("categorize", {
   )
   suppressMessages(
     expect_error(
-      categorize(ex_people[c(1, 1), ], ex_icd10, "elixhauser", id = "name"),
+      categorize(
+        ex_people[c(1, 1), ],
+        codedata = ex_icd10,
+        cc = "elixhauser",
+        id = "name",
+        code = "icd10"
+      ),
       "Non-unique ids!"
     )
 
   )
   expect_equal(
-    ncol(categorize(ex_people, ex_icd10, "elixhauser",
-                    id = "name", index = "index_sid30")),
+    ncol(categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
+                    id = "name", code = "icd10", index = "index_sid30")),
     34
   )
 
   expect_equal(
-    ncol(categorize(ex_people, ex_icd10, "elixhauser", id = "name")),
+    ncol(categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
+                    id = "name", code = "icd10")),
     40
   )
 
   expect_equal(
     names(
-      categorize(ex_people, ex_icd10, "elixhauser", id = "name",
-                    cc_args = list(regex = "regex_icd10", tech_names = TRUE))
+      categorize(
+        ex_people, codedata = ex_icd10, cc = "elixhauser",
+        id = "name", code = "icd10",
+        cc_args = list(regex = "regex_icd10", tech_names = TRUE))
     )[3],
     "elixhauser_regex_icd10_congestive_heart_failure"
   )
 
   expect_equal(
-    nrow(categorize(ex_people, ex_icd10, "elixhauser", id = "name")),
+    nrow(categorize(ex_people, codedata = ex_icd10, cc = "elixhauser",
+                    id = "name", code = "icd10")),
     nrow(ex_people)
   )
 
   expect_warning(
-    categorize(x, "charlson", id = "name"),
+    categorize(x, cc = "charlson"),
     "Output might contain extra columns as left-overs"
   )
 
   expect_identical(
-    suppressWarnings(categorize(x, "charlson", id = "name")[,-"hdia"]),
-    categorize(ex_people, ex_icd10, "charlson", id = "name", codify_args = list(date = "event", days = c(-365, 0)))
+    suppressWarnings(
+      categorize(x, cc = "charlson")[,-c("hdia", "admission", "icd10")]),
+    categorize(
+      ex_people, codedata = ex_icd10, cc = "charlson", id = "name",
+      code = "icd10",
+      codify_args = list(date = "surgery", code_date = "admission", days = c(-365, 0))
+    )
   )
 })
