+++ tests/testthat/test-set_classcodes.R
@@ -1,5 +1,6 @@
 y <- x <-
-  codify(ex_people, ex_icd10, id = "name", date = "event", days = c(-365, 0))
+  codify(ex_people, ex_icd10, id = "name", date = "surgery", code = "icd10",
+         code_date = "admission", days = c(-365, 0))
 
 x <- suppressWarnings(classify(x, "elixhauser"))
 y <- suppressWarnings(classify(y, elixhauser))

[…]
