### 2020-09-24 6a8c37a
Present output as tibble.
Fix #122. Part of #121.
