+++ R/classcodes.R
@@ -72,8 +72,8 @@ as.classcodes <- function(x, hierarchy = attr(x, "hierarchy"), .name = NULL) {
 
   rgs <- colnames(x)[startsWith(colnames(x), "regex")]
   structure(
-    x,
-    class       = unique(c("classcodes", "tbl_df", "tbl", class(x))),
+    tibble::as_tibble(x),
+    class       = unique(c("classcodes", class(x))),
     regexprs    = colnames(x)[startsWith(colnames(x), "regex")],
     indices     = colnames(x)[startsWith(colnames(x), "index")],
     hierarchy   = hierarchy,
