+++ NAMESPACE
@@ -9,12 +9,14 @@ S3method(as.keyvalue,classcodes)
 S3method(categorize,codified)
 S3method(categorize,data.frame)
 S3method(categorize,data.table)
+S3method(categorize,tbl_df)
 S3method(classify,codified)
 S3method(classify,data.frame)
 S3method(classify,data.table)
 S3method(classify,default)
 S3method(index,data.frame)
 S3method(index,matrix)
+S3method(print,codebook)
 S3method(print,summary.classcodes)
 S3method(summary,classcodes)
 S3method(visualize,character)

[…]
