### 2020-09-29 e80abaf
fixed checks
