+++ tests/testthat/test-codebook.R
@@ -6,7 +6,7 @@ cb2 <- codebook(elixhauser, "icd9cmd",
 cbs <- codebooks(cb1 = cb, cb2 = cb2)
 
 test_that("codebook", {
-  expect_is(cb, "list")
+  expect_is(cb, "codebook")
   expect_named(cb, c("readme", "summary", "all_codes"))
   expect_named(
     cbs,
