+++ README.md
@@ -1,25 +1,27 @@
-coder
-================
+---
+title: "coder"
+output: 
+  github_document:
+    df_print: tibble
+---
 
-[![R build
-status](https://github.com/eribul/coder/workflows/R-CMD-check/badge.svg)](https://github.com/eribul/coder/actions)
+[![R build status](https://github.com/eribul/coder/workflows/R-CMD-check/badge.svg)](https://github.com/eribul/coder/actions)
 [![codecov](https://codecov.io/gh/eribul/coder/branch/master/graph/badge.svg)](https://codecov.io/gh/eribul/coder)
-[![Project Status: Active – The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
-<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+<!-- README.md is generated from README.Rmd. Please edit that file --> 
+
+
 
 ## Aim of the package
 
-The goal of `coder` is to classify items from one dataset, using codes
-from a secondary source with classification schemes based on regular
-expressions and weighted indices.
+The goal of `coder` is to classify items from one dataset, using codes from a secondary source with classification schemes based on regular expressions and weighted indices. 
 
-## Installation
 
-You can (soon) install the released version of coder from
-[CRAN](https://CRAN.R-project.org) with:
+## Installation 
+
+You can (soon) install the released version of coder from [CRAN](https://CRAN.R-project.org) with:
 
 ``` r
 # install.packages("coder") # Not yet!
@@ -32,135 +34,71 @@ And the development version from [GitHub](https://github.com/) with:
 remotes::install_github("eribul/coder")
 ```
 
+
 ## Typical use case
 
-**Patient data:** The initial rationale for the package was to classify
-patient data based on medical coding. A typical use case would consider
-patients from a medical/administrative data base, as identified by some
-patient id and possibly with some associated date of interest (date of
-diagnoses/treatment/intervention/hospitalization/rehabilitation). This
-data source could be for example an administrative hospital register or
-a national quality register.
+**Patient data:** The initial rationale for the package was to classify patient data based on medical coding. A typical use case would consider patients from a medical/administrative data base, as identified by some patient id and possibly with some associated date of interest (date of diagnoses/treatment/intervention/hospitalization/rehabilitation). This data source could be for example an administrative hospital register or a national quality register. 
 
-**Codify:** The primary source could then be linked to a secondary
-(possibly larger) data base including the same patients with
-corresponding id:s and some coded patient data. This could be a national
-patient register with medical codes from the International
-Classification of Diseases *(ICD)* with corresponding dates of hospital
-visits/admission/discharge, or a medical prescription register with
-codes from the Anatomic Therapeutic Chemical *(ATC)* classification
-system with dates of medical prescription/dispatch/usage. A time window
-could be specified relating the date of the primary source (i. e. the
-date of a primary total hip arthroplasty; THA), to dates from the
-secondary source (i.e. the date of a medical prescription). ATC codes
-associated with medical prescriptions during one year prior to THA,
-could thus be identified and used as a measure of comorbidity. Another
[…]

[…]
