+++ R/misc.R
@@ -44,7 +44,6 @@ spss2date <- function(x) {
   as.Date(x / 86400, origin = "1582-10-14")
 }
 
-is.Date <- function(x) data.class(x) == "Date"
 
 clean <- function(x) gsub("\\W", "_", tolower(x), perl = TRUE)
 
