### 2020-10-07 915cc3f
fix tibble output in vignettes
