+++ vignettes/classcodes.Rmd
@@ -28,11 +28,9 @@ all_classcodes()$classcodes
 
 We will have a closer look at one of those objects, `charlson`, to exemplify its content. The object is basically a `tibble` (data frame):
 
-
 ```{r}
 charlson
 ```
-
 
 ... with some additional attributes:
 
@@ -42,20 +40,19 @@ attributes(charlson)
 
 # Columns
 
-- The "group" column contains short descriptive names of all groups to classify by. Those names might contain spaces and other special characters which might be transformed when used (argument `tech_names` in `set_classcodes`). 
-- Column "description" (optional) provides additional details of each group. 
-- All columns with names prefixed by `regex_` (also found by `attr(charlson, "regexprs")`) specify different versions of group identification based on regular expressions (see `?charlson` for details). Those regular expressions only identifies individual codes but this can be controlled further by arguments `start` and `stop` to `set_classcodes`. 
-- Columns with names prefixed by `index_` (also found by `attr(charlson, "indices")`) are numeric variables used as weights when calculating index sums based on all (or a subset of) individual groups.
-- Other classcodes objects might have an additional column "condition" for conditional classification (see for example `?hip_ae`).
+-   The "group" column contains short descriptive names of all groups to classify by. Those names might contain spaces and other special characters which might be transformed when used (argument `tech_names` in `set_classcodes`).
+-   Column "description" (optional) provides additional details of each group.
+-   All columns with names prefixed by `regex_` (also found by `attr(charlson, "regexprs")`) specify different versions of group identification based on regular expressions (see `?charlson` for details). Those regular expressions only identifies individual codes but this can be controlled further by arguments `start` and `stop` to `set_classcodes`.
+-   Columns with names prefixed by `index_` (also found by `attr(charlson, "indices")`) are numeric variables used as weights when calculating index sums based on all (or a subset of) individual groups.
+-   Other classcodes objects might have an additional column "condition" for conditional classification (see for example `?hip_ae`).
 
 # Hierarchy
 
-Some classcodes objects (see for example `?elixhauser`) might have an additional class attribute "hierarchy", controlling hierarchical groups where only one of possibly several groups (occurring simultaneity) should be recognized for calculation of a weighted index sum. 
-
+Some classcodes objects (see for example `?elixhauser`) might have an additional class attribute "hierarchy", controlling hierarchical groups where only one of possibly several groups (occurring simultaneity) should be recognized for calculation of a weighted index sum.
 
 # Interpret the regular expressions
 
-## Visualise graphically
+## Visualize graphically
 
 Regular expressions might be difficult to interpret but can be visualized by `visualize()`. This will usually display a web page in the default browser (here suppressed by `show = FALSE`):
 
@@ -63,7 +60,7 @@ Regular expressions might be difficult to interpret but can be visualized by `vi
 url <- visualize(charlson, show = FALSE, regex = "regex_icd10")
 ```
 
-Here, we can open the `url` "manually" by a link to the [JavaScript Regular Expression Parser & Visualizer](`r url`). The displayed visualization is rather complex, however, wherefore a further simplification might be needed before actual understanding of the concept. We can chose to inspect the regular expression for `r charlson$group[1]` as codified by `regex_icd9cm_deyo`:
[…]
