+++ man/print.codebook.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/codebook.R
+\name{print.codebook}
+\alias{print.codebook}
+\title{Print codebook}
+\usage{
+\method{print}{codebook}(x, ...)
+}
+\arguments{
+\item{x}{object of class \link{codebook}}
+
+\item{...}{arguments passed to \code{tibble:::print.tbl()}}
+}
+\value{
+\code{x} (invisible)
+}
+\description{
+Print a preview of sheets otherwise exported to Excel
+}
+\examples{
+codebook(charlson, "icd10cm")
+}
