+++ vignettes/classcodes.Rmd
@@ -26,7 +26,7 @@ Classification is implemented by classcodes objects with regular expressions ide
 all_classcodes()$classcodes
 ```
 
-We will have a closer look at one of those objects, `charlson`, to exemplify its content. The object is basically a `tibble` (data frame):
+Each of those classcodes are documented. See for example `?charlson` for detailed information regarding the implemented classcodes object for the Charlson comorbidity classification. We might also have a closer look at the classcodes itself to exemplify its content. The object is basically a `tibble` (data frame) ...
 
 ```{r}
 charlson
@@ -40,27 +40,35 @@ attributes(charlson)
 
 # Columns
 
--   The "group" column contains short descriptive names of all groups to classify by. Those names might contain spaces and other special characters which might be transformed when used (argument `tech_names` in `set_classcodes`).
--   Column "description" (optional) provides additional details of each group.
--   All columns with names prefixed by `regex_` (also found by `attr(charlson, "regexprs")`) specify different versions of group identification based on regular expressions (see `?charlson` for details). Those regular expressions only identifies individual codes but this can be controlled further by arguments `start` and `stop` to `set_classcodes`.
--   Columns with names prefixed by `index_` (also found by `attr(charlson, "indices")`) are numeric variables used as weights when calculating index sums based on all (or a subset of) individual groups.
--   Other classcodes objects might have an additional column "condition" for conditional classification (see for example `?hip_ae`).
+First, consider the tibble columns, which have pre-specified names and content:
+
+-   `group`: short descriptive names of all groups to classify by (i.e. medical conditions/comorbidities). Might contain spaces and other special characters which might be transformed when used (controlled by `tech_names` in `set_classcodes()`).
+-   `description:` (optional) additional details describing each group.
+-   `regex_*:` regular expressions identifying each group (see `?charlson` for details). Multiple versions might be used if combined with different code sets (i.e. ICD-9 versus ICD-10). (Those column names are also found by `attr(charlson, "regexprs").`)
+-   `index_*`: numeric variables used as weights when calculating index sums based on all (or a subset of) individual groups. (Those column names are also found by `attr(charlson, "indices")` .)
+-   `condition`: (optional; not used with `charlson`) conditional classification (see for example `?hip_ae`).
 
 # Hierarchy
 
-Some classcodes objects (see for example `?elixhauser`) might have an additional class attribute "hierarchy", controlling hierarchical groups where only one of possibly several groups (occurring simultaneity) should be recognized for calculation of a weighted index sum.
+Some classcodes objects have an additional class attribute "hierarchy", controlling hierarchical groups where only one of possibly several groups (occurring simultaneity) should be recognized for calculation of a weighted index sum. This is true for Elixhauser:\
+
+```{r}
+attr(elixhauser, "hierarchy")
+```
+
+... meaning that patients who have both metastatic cancer and solid tumors should be recognized as such. If such patient is later assigned an aggregated index score, however, only the largest score is used (in this case for a metastatic cancer as superior to a solid tumor). The same is true for patients with both uncomplicated and complicated diabetes, for which only the complicated version is recognized.
 
 # Interpret the regular expressions
 
 ## Visualize graphically
 
-Regular expressions might be difficult to interpret but can be visualized by `visualize()`. This will usually display a web page in the default browser (here suppressed by `show = FALSE`):
+Regular expressions might be difficult to interpret and can therefore be `visualize()`:d, This will display a web page in the default browser (here suppressed by `show = FALSE`):
 
 ```{r}
[…]
