+++ man/print.codebook.Rd
@@ -20,3 +20,16 @@ Print a preview of sheets otherwise exported to Excel
 \examples{
 codebook(charlson, "icd10cm")
 }
+\seealso{
+Other classcodes: 
+\code{\link{all_classcodes}()},
+\code{\link{as.data.frame.classified}()},
+\code{\link{classcodes}},
+\code{\link{codebooks}()},
+\code{\link{codebook}()},
+\code{\link{print.summary.classcodes}()},
+\code{\link{set_classcodes}()},
+\code{\link{summary.classcodes}()},
+\code{\link{visualize.classcodes}()}
+}
+\concept{classcodes}
