+++ man/visualize.classcodes.Rd
@@ -67,6 +67,7 @@ Other classcodes:
 \code{\link{classcodes}},
 \code{\link{codebooks}()},
 \code{\link{codebook}()},
+\code{\link{print.codebook}()},
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()}
