+++ man/classcodes.Rd
@@ -77,6 +77,7 @@ Other classcodes:
 \code{\link{as.data.frame.classified}()},
 \code{\link{codebooks}()},
 \code{\link{codebook}()},
+\code{\link{print.codebook}()},
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
@@ -87,6 +88,7 @@ Other classcodes:
 \code{\link{as.data.frame.classified}()},
 \code{\link{codebooks}()},
 \code{\link{codebook}()},
+\code{\link{print.codebook}()},
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
