+++ man/summary.classcodes.Rd
@@ -49,6 +49,7 @@ Other classcodes:
 \code{\link{classcodes}},
 \code{\link{codebooks}()},
 \code{\link{codebook}()},
+\code{\link{print.codebook}()},
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{visualize.classcodes}()}
