### 2020-10-13 ec4d439
codify as S3
Part of #121.
