+++ NAMESPACE
@@ -14,10 +14,13 @@ S3method(classify,codified)
 S3method(classify,data.frame)
 S3method(classify,data.table)
 S3method(classify,default)
+S3method(codify,data.frame)
+S3method(codify,data.table)
 S3method(index,data.frame)
 S3method(index,matrix)
 S3method(print,classified)
 S3method(print,codebook)
+S3method(print,codified)
 S3method(print,summary.classcodes)
 S3method(summary,classcodes)
 S3method(visualize,character)
