+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: coder
 Type: Package
 Title: Deterministic Categorization of Items Based on External Code Data
-Version: 0.11.13
+Version: 0.11.14
 Authors@R: person("Erik", "Bulow", email = "eriklgb@gmail.com",
                   role = c("aut", "cre"),
                   comment = c(ORCID = "0000-0002-9973-456X"))
