### 2020-10-13 1d5a0a9
Change help text
Fix #118.
