### 2020-10-13 b4c3056
Better example data
Fix #126.
