+++ R/manual_for_datasets.R
@@ -324,12 +324,19 @@
 #' Example data for fictive ICD-10-diagnoses to use for testing and
 #' in examples.
 #'
+#' @source
+#'   https://github.com/jackwasey/icd.data
+#'   https://ustur.wsu.edu/about-us/
+#'
 #' @format Data frames with 1,000 rows and 4 variables:
 #' \describe{
 #'   \item{id}{Random names corresponding to column `name` in dataset
 #'     \code{ex_people}}
 #'   \item{date}{random dates corresponding to registered (comorbidity) codes}
-#'   \item{code}{(comorbidity) codes as given by ICD-10}
+#'   \item{code}{ICD-10 codes from the `uranium_pathology`
+#'   dataset in the `icd.data` package by Jack Wasey originating from the
+#'   United States Transuranium and Uranium Registries,
+#'   published in the public domain.}
 #'   \item{hdia}{boolean marker if corresponding code is the main diagnose of
 #'     the hospital visit (randomly assigned to 10 percent of the codes)}
 #' }
