+++ R/misc.R
@@ -86,3 +86,23 @@ decoder_data <- function(x) {
     stop("'coding' should be one of: ", paste(dts, collapse = ", "))
   }
 }
+
+
+
+
+print_tibble <- function(x, ..., n = 10) {
+  if (!is.null(n)) {
+    writeLines(
+      paste0("\nThe printed data is of class: ",
+             paste(class(x), collapse = ", "),
+             ". It has ", nrow(x), " rows.",
+             "\nIt is here previewed as a tibble",
+             "\nUse `print(x, n = NULL)` to print as is ",
+             "(or use `n` to specify the number of rows to preview)!\n\n"
+      )
+    )
+    print(tibble::as_tibble(utils::head(x, n)))
+  } else {
+    NextMethod()
+  }
+}
