+++ data-raw/example_data.R
@@ -1,25 +1,21 @@
 set.seed(123)
 # Generate some example data to include in the package
 
-n <- 1e2     # How many people to look for
-N <- n * 10  # How many cases in the codedata set
+ex_icd10 <-
+  icd.data::uranium_pathology %>%
+  as_tibble() %>%
+  transmute(
+    name = randomNames::randomNames(n())[case],
+    admission = Sys.Date() - sample(0:365, n(), TRUE),
+    icd10     = sample(decoder::icd10se$key, n(), TRUE),
+    hdia      = sample(c(TRUE, FALSE), n(), TRUE, c(.1, .9))
+  )
 
-
-# Individuals we are interested in
 ex_people <-
   tibble::tibble(
-    name    = randomNames::randomNames(n),
-    surgery = Sys.Date() - sample(0:365, n, TRUE)
-)
-
-
-# Codedata for these (and other) individuals
-ex_icd10 <-
-  tibble::tibble(
-    name      = sample(c(ex_people$name, randomNames::randomNames(n)), N, TRUE),
-    admission = Sys.Date() - sample(0:365, N, TRUE) - 100,
-    icd10     = sample(decoder::icd10se$key, N, TRUE),
-    hdia      = sample(c(TRUE, FALSE), N, TRUE, c(.1, .9))
+    name    = sample(unique(ex_icd10$name), 100),
+    surgery = Sys.Date() - sample(0:365, 100, TRUE)
   )
+
 
 usethis::use_data(ex_people, ex_icd10, overwrite = TRUE)
