### 2020-10-13 036f5ce
Update README.Rmd
Fix #114.
