### 2020-10-13 6cdc930
remove (reg|ind)ex_ prefix
Part of #120
