### 2020-10-27 1944cce
bug fix
It was previously a problem if codedata had column named "code_date" since this named masked the externally set name accessed by an object with this name.
