### 2020-10-28 46652c6
Specify regex and indices explicitly to as.classcodes
Fix #130.
