+++ R/classcodes.R
@@ -26,7 +26,10 @@
 #'
 #'
 #' @param x data frame with columns described in the details section
-#' @param hierarchy list of pairwise group names to appear as superior and
+#' @param regex,indices character vector with names of columns in `x` containing
+#'   regular expressions/indices.
+#' @param hierarchy named list of pairwise group names to appear as superior and
+#'   subordinate for indices. See [elixhauser] for a possible application.
 #' @param ... arguments passed between methods
 #' subordinate classes.
 #'   To be used for indexing when the subordinate class is redundant
@@ -54,8 +57,7 @@
 #'
 #' @export
 #' @name classcodes
-#' @examples
-#' as.classcodes(coder::elixhauser)
+#' @example man/examples/as.classcodes.R
 #' @family classcodes
 as.classcodes <- function(x, ...) {
   UseMethod("as.classcodes")
@@ -63,37 +65,69 @@ as.classcodes <- function(x, ...) {
 
 #' @export
 #' @rdname classcodes
-as.classcodes.classcodes <- function(x, ...) {
-  attr(x, "regexprs") <- intersect(attr(x, "regexprs"), names(x))
-  attr(x, "indices") <- intersect(attr(x, "indices"), names(x))
+as.classcodes.classcodes <- function(x,
+                                     ...,
+                                     regex     = attr(x, "regexpr"),
+                                     indices   = attr(x, "indices"),
+                                     hierarchy = attr(x, "hierarchy")
+                                     ) {
+  attr(x, "regexprs")  <- intersect(regex,   names(x))
+  attr(x, "indices")   <- intersect(indices, names(x))
+  attr(x, "hierarchy") <- hierarchy
   check_classcodes(x)
   x
 }
 
 #' @export
 #' @rdname classcodes
-as.classcodes.default <- function(x, ..., hierarchy = attr(x, "hierarchy"), .name = NULL) {
+as.classcodes.default <- function(x,
+                                  ...,
+                                  regex     = NULL,
+                                  indices   = NULL,
+                                  hierarchy = attr(x, "hierarchy"),
+                                  .name     = NULL
+                                  ) {
 
   # To avoid infinite recursive looping due to `$<-.classcodes` method
   class(x) <- setdiff(class(x), "classcodes")
 
-  check_classcodes(x)
+  out <-
+    tibble::as_tibble(
+      stats::setNames(x, gsub("(reg|ind)ex_", "", names(x)))
+    )
 
-
-  rgs  <- colnames(x)[startsWith(colnames(x), "regex_")]
-  indx <- colnames(x)[startsWith(colnames(x), "index_")]
-  names(x) <- gsub("(reg|ind)ex_", "", names(x))
-
-  structure(
-    tibble::as_tibble(x),
-    class       = unique(c("classcodes", class(x))),
-    regexprs    = gsub("regex_", "", rgs),
-    indices     = gsub("index_", "", indx),
-    hierarchy   = hierarchy,
-    name        = .name
-  )
+  out <-
+    structure(
+      out,
+      class       = unique(c("classcodes", class(out))),
+      regexprs    = find_attr(x, regex, "regular expressions", "regex_", TRUE),
+      indices     = find_attr(x, indices, "indices", "index_", FALSE),
+      hierarchy   = hierarchy,
+      name        = .name
+    )
+  check_classcodes(out)
+  out
 }
 
+# Set regex and indices attributes either by arguments,
+# or by prefixed column names
[…]
