+++ R/categorize.R
@@ -117,6 +117,7 @@ categorize.codified <- function(
   cc_args$cc  <- cc_name <- cc
   cc          <- do.call(set_classcodes, cc_args)
   cl          <- classify(codified, cc, cc_args = NULL) # NULL since cc set
+  ...data_cols <- NULL # to avoid check notes
   data        <- unique(codified, by = id)[, ...data_cols]
   data$id_chr <- as.character(data[[id]]) # To be able to merge
   out         <- merge(data, as.data.table(cl),
