+++ man/codebooks.Rd
@@ -33,6 +33,7 @@ Other classcodes:
 \code{\link{as.data.frame.classified}()},
 \code{\link{classcodes}},
 \code{\link{codebook}()},
+\code{\link{print.classcodes}()},
 \code{\link{print.codebook}()},
 \code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},

[…]
