### 2020-10-28 9f946c2
Syntactic names from categorize
Fix #120.
