+++ vignettes/coder.Rmd
@@ -243,7 +243,7 @@ sum(
     cc = rxriskv, 
     id = "name",
     code = "atc"
-  )$index_pratt,
+  )$pratt,
   na.rm = TRUE
 )
 ```
