### 2020-10-28 0c70eba
add examples
Part of #103
