+++ man/all_classcodes.Rd
@@ -20,11 +20,8 @@ all_classcodes()
 Other classcodes: 
 \code{\link{as.data.frame.classified}()},
 \code{\link{classcodes}},
-\code{\link{codebooks}()},
 \code{\link{codebook}()},
 \code{\link{print.classcodes}()},
-\code{\link{print.codebook}()},
-\code{\link{print.summary.classcodes}()},
 \code{\link{set_classcodes}()},
 \code{\link{summary.classcodes}()},
 \code{\link{visualize.classcodes}()}

[…]
