### 2020-11-09 a5f00e6
delete spss2date since no longer used
