+++ tests/testthat/test-misc.R
@@ -3,8 +3,6 @@ context("misc")
 test_that("misc", {
   expect_equal(iris, copybig(iris, FALSE))
 
-  expect_equal(spss2date(0), as.Date("1582-10-14"))
-
   expect_equal(clean_text("prefix", "Hello World!"), "prefix_hello_world_")
   expect_error(
     clean_text(iris, "wrong"),
