+++ man/spss2date.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/misc.R
-\name{spss2date}
-\alias{spss2date}
-\title{Convert SPSS date to valid date}
-\usage{
-spss2date(x)
-}
-\arguments{
-\item{x}{SPSS date}
-}
-\value{
-Date
-}
-\description{
-Dates from SPSS are stored as seconds from 1582-10-14.
-We convert to days from 1970-01-01
-}
-\seealso{
-Other helper: 
-\code{\link{as.keyvalue.classcodes}()}
-}
-\concept{helper}
-\keyword{internal}
