### 2020-11-09 9277940
mixed
Including some examples as part of #103.
Also includes a hex sticker logo
