+++ tests/testthat/test-index.R
@@ -1,4 +1,3 @@
-context("index")
 y <-
   codify(ex_people, ex_icd10, id = "name", code = "icd10",
     date = "surgery", code_date = "admission", days = c(-365, 0)) %>%
