+++ README.Rmd
@@ -124,7 +124,12 @@ sum(ch$malignancy, na.rm = TRUE)
 
 Distribution of the index as proposed by QUan et al 2011 during the 90 day period:
 ```{r}
-hist(ch$quan_updated, main = "COmorbidity index by Quan et al. 2011")
+hist(
+  ch$quan_updated, 
+  col = "lightblue",
+  main = "Charlson comorbidity",
+  xlab = "Comorbidity index by Quan et al. 2011"
+)
 ```
 
 
