### 2020-11-17 ccbe64a
Additional examples
Fix #103
