+++ vignettes/classcodes.Rmd
@@ -7,6 +7,7 @@ vignette: >
   %\VignetteIndexEntry{Classcodes objects}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+bibliography: references.bib
 ---
 
 ```{r, include = FALSE}
@@ -46,11 +47,11 @@ categorize(
 )
 ```
 
-Here, `charlson` (as supplied by the `cc` argument) is a "classcodes" object containing a classification scheme matching patient codes (from `codedata`) to groups (comorbidity conditions) by regular expressions.
+Here, `charlson` (as supplied by the `cc` argument) is a "classcodes" object containing a classification scheme. This is the specification of how to match patient codes (from `codedata`) to groups (comorbidity conditions). It is based on regular expressions which recognize all codes for each group.
 
 # Default classcodes
 
-There are `r nrow(all_classcodes())` default "classcodes" objects in the package (`classcodes` column below). Each of them might have several versions of regular expressions (`regex`) and weighted indices (`indices`):
+There are `r nrow(all_classcodes())` default "classcodes" objects in the package (`classcodes` column below). Each of them might have several versions of regular expressions (column `regex`) and weighted indices (column `indices`):
 
 ```{r}
 all_classcodes()
@@ -58,49 +59,48 @@ all_classcodes()
 
 # classcodes object
 
-Each of those classcodes are documented (see for example `?charlson`). The object is basically a `tibble` (data frame) with some additional attributes (and a print method to present those):
+Each of those classcodes objects are documented (see for example `?charlson`). The object is basically a `tibble` (data frame) with some additional attributes (and a print method to present those):
 
 ```{r}
 charlson
 ```
 
-Columns have pre-specified names and content:
+Columns have pre-specified names and/or content:
 
 -   `group`: short descriptive names of all groups to classify by (i.e. medical conditions/comorbidities).
 -   `description:` (optional) additional details describing each group.
 -   regular expressions identifying each group (see `?charlson` for details). Multiple versions might be used if combined with different code sets (i.e. ICD-9 versus ICD-10). (Column names are arbitrary but identified by `attr(., "regexprs").`)
--   numeric variables used as weights when calculating index sums based on all (or a subset of) individual groups. (Column names are arbitrary but identified by `attr(., "indices")` .)
+-   numeric vectors used as weights when calculating index sums based on all (or a subset of) individual groups. (Column names are arbitrary but identified by `attr(., "indices")`.)
 -   `condition`: (optional; not used with `charlson`) conditional classification (see for example `?hip_ae`).
 
-In the example above, we did not specify which version of regular expression to use. We see from the printed output above (or by `attr(charlson, "regexprs")`), however, that the first listed regular expression (which is used by default) is "icd10". We have ICD-10 codes recorded in our code data set (`ex_icd10`). We could therefore use the alternative "icd10\_rcs" as well. Other versions might be relevant if the medical data is coded by other classifications (such as earlier versions of ICD). We will show below how to alter this setting in practice.
+In the example above, we did not specify which version of regular expression to use. We see from the printed output above (or by `attr(charlson, "regexprs")`), that the first regular expression is "icd10". This will be used by default. We have ICD-10 codes recorded in our code data set (`ex_icd10$icd10`). We might therefore use either "icd10" or "icd10\_rcs" to classify those. Other versions might be relevant if the medical data is coded by other codes (such as earlier versions of ICD). We will show below how to alter this setting in practice.
 
 # Hierarchy
 
[…]
