+++ vignettes/references.bib
@@ -1,0 +1,15 @@
+
+@article{Quan2005,
+	title = {Coding algorithms for defining comorbidities in ICD-9-CM and ICD-10 administrative data.},
+	author = {{Quan}, {Hude} and {Sundararajan}, {Vijaya} and {Halfon}, {Patricia} and {Fong}, {Andrew} and {Burnand}, {Bernard} and {Luthi}, {Jean-Christophe} and {Saunders}, {L Duncan} and {Beck}, {Cynthia a} and {Feasby}, {Thomas E} and {Ghali}, {William a}},
+	year = {2005},
+	date = {2005},
+	journal = {Medical care},
+	pages = {1130--1139},
+	volume = {43},
+	number = {11},
+	doi = {10.1097/01.mlr.0000182534.19832.83},
+	note = {Citation Key: Quan2005
+PMID: 16224307
+tex.isbn: 0025-7079 (Print) \r0025-7079 (Linking)}
+}
