+++ vignettes/references.bib
@@ -13,3 +13,136 @@
 PMID: 16224307
 tex.isbn: 0025-7079 (Print) \r0025-7079 (Linking)}
 }
+
+@article{Deyo1992,
+	title = {Adapting a clinical comorbidity index for use with ICD-9-CM administrative databases},
+	author = {{Deyo}, {Richard A.} and {Cherkin}, {Daniel C.} and {Ciol}, {Marcia A.}},
+	year = {1992},
+	month = {06},
+	date = {1992-06},
+	journal = {Journal of Clinical Epidemiology},
+	pages = {613--619},
+	volume = {45},
+	number = {6},
+	doi = {10.1016/0895-4356(92)90133-8},
+	url = {https://www-sciencedirect-com.ezproxy.ub.gu.se/science/article/pii/0895435692901338},
+	note = {Citation Key: Deyo1992
+PMID: 1607900
+tex.isbn: ISSN  0895-4356
+tex.publisher: Pergamon}
+}
+
+@article{Brusselaers2017,
+	title = {The Charlson comorbidity index in registry-based research},
+	author = {{Brusselaers}, {Nele} and {Lagergren}, {Jesper}},
+	year = {2017},
+	month = {01},
+	date = {2017-01},
+	journal = {Methods of Information in Medicine},
+	pages = {401--406},
+	volume = {56},
+	number = {05},
+	doi = {10.3414/ME17-01-0051},
+	url = {http://www.thieme-connect.de/DOI/DOI?10.3414/ME17-01-0051},
+	note = {Citation Key: Brusselaers2017}
+}
+
+@article{Quan2005,
+	title = {Coding algorithms for defining comorbidities in ICD-9-CM and ICD-10 administrative data.},
+	author = {{Quan}, {Hude} and {Sundararajan}, {Vijaya} and {Halfon}, {Patricia} and {Fong}, {Andrew} and {Burnand}, {Bernard} and {Luthi}, {Jean-Christophe} and {Saunders}, {L Duncan} and {Beck}, {Cynthia a} and {Feasby}, {Thomas E} and {Ghali}, {William a}},
+	year = {2005},
+	date = {2005},
+	journal = {Medical care},
+	pages = {1130--1139},
+	volume = {43},
+	number = {11},
+	doi = {10.1097/01.mlr.0000182534.19832.83},
+	note = {Citation Key: Quan2005
+PMID: 16224307
+tex.isbn: 0025-7079 (Print) \r0025-7079 (Linking)}
+}
+
+@article{Pratt2018,
+	title = {The validity of the Rx-Risk comorbidity index using medicines mapped to the anatomical therapeutic chemical (ATC) classification system},
+	author = {{Pratt}, {Nicole L.} and {Kerr}, {Mhairi} and {Barratt}, {John D.} and {Kemp-Casey}, {Anna} and {Ellett}, {Lisa M. Kalisch} and {Ramsay}, {Emmae} and {Roughead}, {Elizabeth Ellen}},
+	year = {2018},
+	month = {04},
+	date = {2018-04},
+	journal = {BMJ Open},
+	volume = {8},
+	number = {4},
+	doi = {10.1136/bmjopen-2017-021122},
+	note = {Citation Key: Pratt2018
+tex.publisher: BMJ Publishing Group}
+}
+
+@article{Feinstein1970,
+	title = {The pre-therapeutic classification of co-morbidity in chronic disease},
+	author = {{Feinstein}, {Alvan R}},
+	year = {1970},
+	date = {1970},
+	journal = {Journal of Chronic Diseases},
+	pages = {455--468},
+	volume = {23},
+	number = {7},
+	doi = {https://doi.org/10.1016/0021-9681(70)90054-8},
[…]
