### 2020-11-19 9c216c1
improved documentation
