+++ R/all_classcodes.R
@@ -1,5 +1,8 @@
 #' Summary data for all default classcodes object in the package
 #'
+#' Tabulate object names and list all related versions of implemented regular expressions
+#' and index weights.
+#'
 #' @return [tibble::tibble] with columns describing all default classcodes
 #'   objects from the package.
 #' @export
