### 2020-11-20 ee275da
increased no of unit tests
Fix #129
