+++ tests/testthat/test-set_classcodes.R
@@ -40,5 +40,16 @@ test_that("set_classcodes", {
     set_classcodes("elixhauser", start = FALSE, stop = TRUE)$icd10[1],
     "(I(099|1(10|3[02])|255|4(2[05-9]|3)|50)|P290)$"
   )
+
+  ch <- as.data.frame(charlson)
+  expect_error(set_classcodes(ch), "cc is not a classcodes object")
+
+  expect_error(
+    set_classcodes(
+      set_classcodes("charlson", tech_names = TRUE),
+      tech_names = FALSE
+    ),
+    "classcodes object has technical names. Either re-specify"
+  )
 })
 
