+++ R/codify.R
@@ -79,12 +79,13 @@ codify <- function(x,
   if (usedate) {
     if (is.null(date))
       stop("Argument 'date' must be specified if 'days' is not NULL!")
-    if (!inherits(x[[date]], "Date"))
-      stop("data column '", date, "' is not of class 'Date'!")
+    if (!date %in% names(x) || !inherits(x[[date]], "Date"))
+      stop(date, " is not a `Date` column of 'x'!")
     if (is.null(code_date))
       stop("Argument 'code_date' must be specified if 'days' is not NULL!")
-    if (!inherits(codedata[[code_date]], "Date"))
-      stop("codedata column '", date, "' is not of class 'Date'!")
+    if (!code_date %in% names(codedata) ||
+        !inherits(codedata[[code_date]], "Date"))
+      stop(code_date, "' is not a `Date` column in 'codedata'!")
   }
   if (!usedate && !is.null(date)) {
     warning("Date column ignored since days = NULL!")
