+++ tests/testthat/test-misc.R
@@ -9,5 +9,11 @@ test_that("misc", {
 
   expect_equal(decoder_data("ben"), decoder::ben)
   expect_error(decoder_data("wrong_yee!"), "'coding' should be one of")
+
+  expect_output(print_tibble(iris), "The printed data is of class: data.frame.")
+  expect_output(
+    print_tibble(iris, n = NULL),
+    "[^(The printed data is of class: data.frame.)]"
+  )
 })
 
