+++ _pkgdown.yml
@@ -1,8 +1,5 @@
 destination: docs
 url: eribul.github.io/coder
-template:
-  params:
-    bootswatch: united
 
 authors:
   Erik Bulow:
