### 2020-11-20 ropensci/coder#89 @eribul
All issues adressed and formal answer send!
