+++ man/visualize.classcodes.Rd
@@ -10,8 +10,8 @@
 \item{x}{\link{classcodes} object or name of such object included in the package
 (see \code{\link[=all_classcodes]{all_classcodes()}}).}
 
-\item{group}{names (as character vector) of groups to visualize (subset of \code{rownames(x)}).
-(All groups if \code{NULL}.)}
+\item{group}{names (as character vector) of groups to visualize
+(subset of \code{rownames(x)}). (All groups if \code{NULL}.)}
 
 \item{show}{should a visualization be shown in the default web browser.
 Set to \code{FALSE} to just retrieve a URL for later use.}
