### 2020-11-24 #381 @zabore
@eribul You have clearly done a lot of work on this package in response to the reviews - well done! The changes you have made to the README, vignettes, and documentation in particular have really improved the package. I am satisfied with your responses to all of my comments, and don't have any further comments.
