### 2020-12-15 6637b16
Fixing a few small typos
