+++ R/manual_for_datasets.R
@@ -312,8 +312,8 @@
 #'   \item{name}{random person names}
 #'   \item{atc}{Random codes from the Anatomic Therapeutic Chemical
 #'   classification (ATC) system.}
-#'   \item{prescription}{random dates of prescription of medications with c
-#'   oresponding ATC codes}
+#'   \item{prescription}{random dates of prescription of medications with
+#'   corresponding ATC codes}
 #' }
 #'
 #' @family example data
