### 2020-12-15 #381 @dgrtwo
Thanks for your very hard work on an already strong package!

I'm especially impressed by the README and vignettes. Both now include several practical examples, and I think the visualizations and "stories" around them make it really clear how the package can be used. I also appreciate that you agreed with and implemented many of my suggestions (as.codedata, returning a tibble, etc). I'm certainly satisfied with the package being added to ROpenSci.

A few remaining thoughts:

* I found a handful of trivial typos in the package and have opened a PR: https://github.com/eribul/coder/pull/134 For that I found `devtools::spell_check()` useful 😃
* You have pkgdown in the repo but I can't find the packagedown site online, is it published? Incidentally, is it normal that the pkgdown docs still have docs for the deleted functions? https://github.com/eribul/coder/blob/master/docs/reference/filter_dates.html

And a few that don't require any action or response:

* `codify` returning a data.table: If I were implementing it I think I would instead have had the following step, `classify`, turn the the table into a data.table (I think the table->data.table transition would be fast on even a large table), but your compromise is still entirely reasonable.
* I'm convinced about the naming conventions for classify, index, etc
* As the original author of "broom" I definitely appreciate the dilemma of creating a package that doesn't have an easily Googleable name, which is why I was recommending against "sharing my fate" 😉 but it's perfectly reasonable to keep the name.
