### 2020-12-16 d136977
reset pkgdown files
Try to solve #136
