+++ .github/workflows/pkgdown.yaml
@@ -44,5 +44,5 @@ jobs:
 
 
       - name: Deploy package
-        run: pkgdown::deploy_to_branch(new_process = FALSE)
+        run: pkgdown::deploy_to_branch(clean = TRUE, new_process = FALSE)
         shell: Rscript {0}
