### 2020-12-16 8c9be67
Updated workflows
Try to solve #137
