### 2020-12-16 #381 @eribul
@zabore and @dgrtwo I would very much like to acknowledge your work as reviewers by the `rev`-role in the description file:
```
person("Emely C", "Zabore", role = "rev",
    comment = "Bea reviewed the package (v. 0.12.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/381>"),
person("David", "Robinson", role = "rev",
    comment = "Bea reviewed the package (v. 0.12.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/381>"),
```
Do I have your permissin to do so? :-)
