### 2020-12-16 #381 @eribul
@noamross I have transferred the repo to ropensci but might have missed to select the "coder" team. It seems then I have not access to the usual "settings" for the repo, wherefore I can not "replace [my] package docs URL with https://docs.ropensci.org/package_name". Is it possible to get the relevant access in hindsight?
