+++ README.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.path = "man/figures/README",
   out.width = "100%"
-)
+) 
 ```
 
 ## Aim of the package
