+++ README.md
@@ -34,8 +34,8 @@ remotes::install_github("eribul/coder")
 
 ## Typical use case
 
-  - Determining comorbidities before clinical trials
-  - Discovering adverse events after surgery
+-   Determining comorbidities before clinical trials
+-   Discovering adverse events after surgery
 
 **Patient data:** The initial rationale for the package was to classify
 patient data based on medical coding. A typical use case would consider
@@ -120,20 +120,20 @@ Assume we have some patients with surgery at specified dates:
 ``` r
 library(coder)
 ex_people
-#> [90m# A tibble: 100 x 2[39m
+#> # A tibble: 100 x 2
 #>    name              surgery   
-#>    [3m[90m<chr>[39m[23m             [3m[90m<date>[39m[23m    
-#> [90m 1[39m Chen, Trevor      2020-09-29
-#> [90m 2[39m Graves, Acineth   2020-06-21
-#> [90m 3[39m Trujillo, Yanelly 2020-06-08
-#> [90m 4[39m Simpson, Kenneth  2020-09-10
-#> [90m 5[39m Chin, Nelson      2020-08-24
-#> [90m 6[39m Le, Christina     2020-03-28
-#> [90m 7[39m Kang, Xuan        2020-06-30
-#> [90m 8[39m Shuemaker, Lauren 2020-03-29
-#> [90m 9[39m Boucher, Teresa   2020-09-04
-#> [90m10[39m Le, Soraiya       2020-08-09
-#> [90m# … with 90 more rows[39m
+#>    <chr>             <date>    
+#>  1 Chen, Trevor      2020-09-29
+#>  2 Graves, Acineth   2020-06-21
+#>  3 Trujillo, Yanelly 2020-06-08
+#>  4 Simpson, Kenneth  2020-09-10
+#>  5 Chin, Nelson      2020-08-24
+#>  6 Le, Christina     2020-03-28
+#>  7 Kang, Xuan        2020-06-30
+#>  8 Shuemaker, Lauren 2020-03-29
+#>  9 Boucher, Teresa   2020-09-04
+#> 10 Le, Soraiya       2020-08-09
+#> # … with 90 more rows
 ```
 
 Those patients (among others) were also recorded in a national patient
@@ -142,20 +142,20 @@ the International Classification of Diseases (ICD) version 10:
 
 ``` r
 ex_icd10
-#> [90m# A tibble: 2,376 x 4[39m
+#> # A tibble: 2,376 x 4
 #>    name                 admission  icd10 hdia 
-#>    [3m[90m<chr>[39m[23m                [3m[90m<date>[39m[23m     [3m[90m<chr>[39m[23m [3m[90m<lgl>[39m[23m
-#> [90m 1[39m Tran, Kenneth        2020-04-12 S134A FALSE
-#> [90m 2[39m Tran, Kenneth        2020-09-26 W3319 FALSE
-#> [90m 3[39m Tran, Kenneth        2020-09-05 Y0262 TRUE 
-#> [90m 4[39m Tran, Kenneth        2020-07-29 X0488 FALSE
-#> [90m 5[39m Sommerville, Dominic 2020-09-17 V8104 FALSE
-#> [90m 6[39m Sommerville, Dominic 2020-04-28 B853  FALSE
-#> [90m 7[39m Sommerville, Dominic 2020-09-12 Q174  FALSE
-#> [90m 8[39m Sommerville, Dominic 2020-05-03 A227  FALSE
-#> [90m 9[39m Sommerville, Dominic 2020-09-07 H702  FALSE
-#> [90m10[39m Sommerville, Dominic 2019-12-31 X6051 TRUE 
-#> [90m# … with 2,366 more rows[39m
[…]
