### 2020-12-16 ropensci/coder#140 @eribul
Submit to JOSS
Follow instructions here: https://github.com/ropensci/software-review/issues/381

Requires repo access, which I currently not have.
