### 2020-12-16 #381 @stefaniebutland
@eribul Congratulations on coder passing peer review! Would love to have a post.
We have openings for publication in late January. Please suggest a date for submitting a draft via pull request and I'll provide a publication date. https://blogguide.ropensci.org/ gives content and technical guidelines.
