### 2020-12-16 #381 @stefaniebutland
@eribul in your post, you should talk about where these codes/categorizations come from, what they mean and how they are used for a general open science that may not be familiar with medical/clinical data, and how his package solves a problem researchers have with using them.
