+++ paper/paper.md
@@ -10,17 +10,17 @@ authors:
     orcid: 0000-0002-9973-456X
     affiliation: "1, 2"
 affiliations:
- - name: The Swedish Hip Arthroplasty Register, Registercentrum Västra Götaland, Gothenburg, Sweden
+ - name: The Swedish Arthroplasty Register, Registercentrum Västra Götaland, Gothenburg, Sweden
    index: 1
  - name: Department of orthopaedics, institute of clinical sciences, Sahlgrenska Academy, University of Gothenburg, Gothenburg, Sweden
    index: 2
-date: 9 August 2018
+date: 16 December 2020
 bibliography: paper.bib
 ---
 
 # Summary
 
-The `coder` package lets researchers classify and categorize coded item data using prespecified classification schemes based on regular expressions. Default classification schemes are included for commonly used medical and clinical classifications. The package implemantation aim for high performance and the package can be used with large data sets.
+The `coder` package lets researchers classify and categorize coded item data using pre-specified classification schemes based on regular expressions. Default classification schemes are included for commonly used medical and clinical classifications. The package implementation aim for high performance and the package can be used with large data sets.
 
 
 # Medical coding and classifications
@@ -34,12 +34,12 @@ Codes and classifications are maintained and developed by several international 
 
 Common classifications such as the International Classification of Diseases (ICD) or the Anatomical Therapeutic Chemical Classification System (ATC) entails thousands of codes which are hard to use and interpret in applied research. This is often solved by an abstraction layer combining individual codes into broader categories, sometimes further simplified by a single index value based on a weighted sum of individual categories [@Charlson1987; @Elixhauser1998; @Quan2005; @Sloan2003; @Pratt2018]. 
  
-Large and long-standing national databases often contain millions of entries and span several Gigabytes (GB) in size. This leads to high computational burden and a time-consuming data managing process, a cumbersome but necessary prerequisite before any relevant analysis can be performed. There are several R-packages with a delibarate focus on comorbidity data coded by ICD and summarized by the Charlson or Elixhauser comorbidity indices ([icd](https://jackwasey.github.io/icd), [comorbidity](https://ellessenne.github.io/comorbidity/) [@Gasparini2018] and [medicalrisk](https://github.com/patrickmdnet/medicalrisk)). The `coder` package includes such capabilities as well, but takes a more general approach to deterministic item classification and categorization.
+Large and long-standing national databases often contain millions of entries and span several Gigabytes (GB) in size. This leads to high computational burden and a time-consuming data managing process, a cumbersome but necessary prerequisite before any relevant analysis can be performed. There are several R-packages with a deliberate focus on comorbidity data coded by ICD and summarized by the Charlson or Elixhauser comorbidity indices ([icd](https://jackwasey.github.io/icd), [comorbidity](https://ellessenne.github.io/comorbidity/) [@Gasparini2018] and [medicalrisk](https://github.com/patrickmdnet/medicalrisk)). The `coder` package includes such capabilities as well, but takes a more general approach to deterministic item classification and categorization.
 
 
 # The coder package
 
[…]
