+++ paper/paper.bib
@@ -1,8 +1,8 @@
 
 @article{Berg2018,
-  title = {No Increase in Readmissions or Adverse Events after Implementation of Fast-Track Program in Total Hip and Knee Replacement at 8 {{Swedish}} Hospitals: {{An}} Observational before-and-after Study of 14,148 Total Joint Replacements 2011\textendash{}2015},
+  title = {No Increase in Readmissions or Adverse Events after Implementation of Fast-Track Program in Total Hip and Knee Replacement at 8 {{Swedish}} Hospitals: {{An}} Observational before-and-after Study of 14,148 Total Joint Replacements 2011--2015},
   shorttitle = {No Increase in Readmissions or Adverse Events after Implementation of Fast-Track Program in Total Hip and Knee Replacement at 8 {{Swedish}} Hospitals},
-  author = {Berg, Urban and B{\"u}Low, Erik and Sundberg, Martin and Rolfson, Ola},
+  author = {Berg, Urban and B{\"u}low, Erik and Sundberg, Martin and Rolfson, Ola},
   year = {2018},
   month = sep,
   volume = {89},
@@ -47,7 +47,7 @@
 }
 
 @article{Bulow2020,
-  title = {Are the First or the Second Hips of Staged Bilateral {{THAs}} More Similar to Unilateral Procedures? {{A}} Study from the Swedish Hip Arthroplasty Register},
+  title = {Are the First or the Second Hips of Staged Bilateral {{THAs}} More Similar to Unilateral Procedures? {{A}} Study from the {{Swedish}} Hip Arthroplasty Register},
   author = {B{\"u}low, Erik and Nemes, Szilard and Rolfson, Ola},
   year = {2020},
   month = mar,
@@ -117,7 +117,7 @@
 }
 
 @article{Gasparini2018,
-  title = {Comorbidity: {{An}} r Package for Computing Comorbidity Scores Software \textbullet{} Review \textbullet{} Repository \textbullet{} Archive},
+  title = {Comorbidity: {{An}} {{R}} Package for Computing Comorbidity Scores Software \textbullet{} Review \textbullet{} Repository \textbullet{} Archive},
   author = {Gasparini, Alessandro},
   year = {2018},
   doi = {10.21105/joss.00648},
@@ -143,7 +143,7 @@
 }
 
 @article{Jawad2019,
-  title = {Multi-State Analysis of Hemi- and Total Hip Arthroplasty for Hip Fractures in the {{Swedish}} Population\textemdash{{Results}} from a {{Swedish}} National Database Study of 38,912 Patients},
+  title = {Multi-State Analysis of Hemi- and Total Hip Arthroplasty for Hip Fractures in the {{Swedish}} Population - {{Results}} from a {{Swedish}} National Database Study of 38,912 Patients},
   author = {Jawad, Z. and Nemes, S. and B{\"u}low, E. and Rogmark, C. and Cnudde, P.},
   year = {2019},
   month = feb,
@@ -176,7 +176,7 @@
 }
 
 @article{Pratt2018,
-  title = {The Validity of the Rx-Risk Comorbidity Index Using Medicines Mapped to the Anatomical Therapeutic Chemical ({{ATC}}) Classification System},
+  title = {The Validity of the {{Rx-Risk}} Comorbidity Index Using Medicines Mapped to the Anatomical Therapeutic Chemical ({{ATC}}) Classification System},
   author = {Pratt, Nicole L. and Kerr, Mhairi and Barratt, John D. and {Kemp-Casey}, Anna and Ellett, Lisa M. Kalisch and Ramsay, Emmae and Roughead, Elizabeth Ellen},
   year = {2018},
[…]
