+++ README.md
@@ -8,6 +8,7 @@ status](https://github.com/ropensci/coder/workflows/R-CMD-check/badge.svg)](http
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![DOI](https://zenodo.org/badge/65975808.svg)](https://zenodo.org/badge/latestdoi/65975808)
+[![status](https://joss.theoj.org/papers/49d52b062aa289d7b990b1a0a678d109/status.svg)](https://joss.theoj.org/papers/49d52b062aa289d7b990b1a0a678d109)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -35,8 +36,8 @@ remotes::install_github("eribul/coder")
 
 ## Typical use case
 
-  - Determining comorbidities before clinical trials
-  - Discovering adverse events after surgery
+-   Determining comorbidities before clinical trials
+-   Discovering adverse events after surgery
 
 **Patient data:** The initial rationale for the package was to classify
 patient data based on medical coding. A typical use case would consider
@@ -134,7 +135,7 @@ ex_people
 #>  8 Shuemaker, Lauren 2020-03-29
 #>  9 Boucher, Teresa   2020-09-04
 #> 10 Le, Soraiya       2020-08-09
-#> # ... with 90 more rows
+#> # … with 90 more rows
 ```
 
 Those patients (among others) were also recorded in a national patient
@@ -156,7 +157,7 @@ ex_icd10
 #>  8 Sommerville, Dominic 2020-05-03 A227  FALSE
 #>  9 Sommerville, Dominic 2020-09-07 H702  FALSE
 #> 10 Sommerville, Dominic 2019-12-31 X6051 TRUE 
-#> # ... with 2,366 more rows
+#> # … with 2,366 more rows
 ```
 
 Using those two data sets, as well as a classification scheme
@@ -175,26 +176,26 @@ ch <-
 
 ch
 #> # A tibble: 100 x 25
-#>    name  surgery    `myocardial inf~ `congestive hea~ `peripheral vas~
+#>    name  surgery    myocardial.infa… congestive.hear… peripheral.vasc…
 #>    <chr> <date>     <lgl>            <lgl>            <lgl>           
-#>  1 Arch~ 2020-05-25 FALSE            FALSE            FALSE           
-#>  2 Awtr~ 2020-09-20 FALSE            FALSE            FALSE           
-#>  3 Bamm~ 2020-03-25 FALSE            FALSE            FALSE           
-#>  4 Bank~ 2020-06-18 FALSE            FALSE            FALSE           
-#>  5 Bega~ 2020-06-19 FALSE            FALSE            FALSE           
-#>  6 Beni~ 2020-02-24 FALSE            FALSE            FALSE           
-#>  7 Bhat~ 2020-10-12 FALSE            FALSE            FALSE           
-#>  8 Bouc~ 2020-09-04 FALSE            FALSE            TRUE            
-#>  9 Carn~ 2020-09-11 FALSE            FALSE            FALSE           
-#> 10 Chap~ 2020-05-18 FALSE            FALSE            FALSE           
-#> # ... with 90 more rows, and 20 more variables: `cerebrovascular
-#> #   disease` <lgl>, dementia <lgl>, `chronic pulmonary disease` <lgl>,
-#> #   `rheumatic disease` <lgl>, `peptic ulcer disease` <lgl>, `mild liver
-#> #   disease` <lgl>, `diabetes without complication` <lgl>, `hemiplegia or
-#> #   paraplegia` <lgl>, `renal disease` <lgl>, `diabetes complication` <lgl>,
[…]
