+++ README.Rmd
@@ -5,6 +5,7 @@ output: github_document
 # coder <img src="man/figures/logo.png" align="right"/>
 
 [![R build status](https://github.com/ropensci/coder/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/coder/actions) [![codecov](https://codecov.io/gh/ropensci/coder/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/coder) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![DOI](https://zenodo.org/badge/65975808.svg)](https://zenodo.org/badge/latestdoi/65975808)
+[![status](https://joss.theoj.org/papers/49d52b062aa289d7b990b1a0a678d109/status.svg)](https://joss.theoj.org/papers/49d52b062aa289d7b990b1a0a678d109)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
