### 2020-12-17 #381 @stefaniebutland
> sound like a great outline
Noam's idea :-)

Draft by 2021-01-25 is great. My colleague @steffilazerte will review it.
