### 2020-12-18 c2a3b33
include DOI for Elixhauser
