+++ paper/paper.bib
@@ -112,6 +112,7 @@
   volume = {36},
   pages = {8--27},
   journal = {Medical care},
+  doi = {10.1097/00005650-199801000-00004},
   keywords = {\#nosource,coder},
   number = {1}
 }
