+++ cran-comments.md
@@ -1,0 +1,7 @@
+Dear CRAN maintainer!
+
+There is a note concerning undefined global variables "...data_cols" and "..code". Those are used as by design for the data.table package.
+I hope that such usage is accaptable.
+
+Best regards!
+Erik Bülow
