### 2021-01-12 d7aa75e
fixed CRAN notes
Fix #144
