+++ man/coder-package.Rd
@@ -10,8 +10,8 @@
 
 Fast categorization of items based on external code data identified by 
   regular expressions. A typical use case considers patient with medically coded 
-  data, such as codes from the International Classification of Diseases (ICD) or 
-  the Anatomic Therapeutic Chemical (ATC) classification system. 
+  data, such as codes from the International Classification of Diseases ('ICD') or 
+  the Anatomic Therapeutic Chemical ('ATC') classification system. 
   Functions of the package relies on a triad of objects: (1) case data with unit 
   id:s and possible dates of interest; (2) external code data for corresponding 
   units in (1) and with optional dates of interest and; (3) a classification 
@@ -19,14 +19,14 @@ Fast categorization of items based on external code data identified by
   categorize relevant codes from (2). 
   It is easy to introduce new classification schemes ('classcodes' objects) or  
   to use default schemes included in the package. Use cases includes patient 
-  categorization based on comorbidity indices such as Charlson, Elixhauser, 
-  RxRisk V, or the co-morbidity-polypharmacy score (CPS), as well as adverse 
-  events after hip and knee arthroplasty surgery.
+  categorization based on 'comorbidity indices' such as 'Charlson', 'Elixhauser', 
+  'RxRisk V', or the 'comorbidity-polypharmacy' score (CPS), as well as adverse 
+  events after hip and knee replacement surgery.
 }
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://docs.ropensci.org/coder}
+  \item \url{https://docs.ropensci.org/coder/}
   \item Report bugs at \url{https://github.com/ropensci/coder/issues}
 }
 

[…]
