### 2021-01-13 f5a3e13
fixmanuak CRAN notes
