### 2021-01-14 a409d9f
CRAN check
fix #145
