+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: coder
 Type: Package
 Title: Deterministic Categorization of Items Based on External Code Data
-Version: 0.13.4
+Version: 0.13.5
 Authors@R: 
   c(
     person("Erik", "Bulow", email = "eriklgb@gmail.com",
