### 2021-01-18 f3d6ecf
added T846F to hip_ae_hailer
Fix #69
