### 2021-01-18 dabd954
bring back examples
