### 2021-01-18 1cceee9
add CRAN info
Fix #32
