+++ README.Rmd
@@ -5,6 +5,8 @@ output: github_document
 # coder <img src="man/figures/logo.png" align="right"/>
 
 [![R build status](https://github.com/ropensci/coder/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/coder/actions) [![codecov](https://codecov.io/gh/ropensci/coder/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/coder) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![DOI](https://zenodo.org/badge/65975808.svg)](https://zenodo.org/badge/latestdoi/65975808) [![status](https://joss.theoj.org/papers/10.21105/joss.02916/status.svg)](https://joss.theoj.org/papers/10.21105/joss.02916)
+[![CRAN status](https://www.r-pkg.org/badges/version/coder)](https://CRAN.R-project.org/package=coder)
+![CRAN downloads](http://cranlogs.r-pkg.org/badges/grand-total/coder)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -23,10 +25,10 @@ The goal of `{coder}` is to classify items from one dataset, using codes from a 
 
 ## Installation
 
-You can (soon) install the released version of coder from [CRAN](https://CRAN.R-project.org) with:
+You can install the released version of coder from [CRAN](https://CRAN.R-project.org) with:
 
 ``` r
-# install.packages("coder") # Not yet!
+# install.packages("coder")
 ```
 
 And the development version from [GitHub](https://github.com/) with:
