### 2021-01-26 fb60faa
reinsert hex logo
Fix #146
